Do not skip projects if a fast-forward merge happens
I noticed that projects.gedmin.as was not listing mgp2pdf.  Apparently
this is because I merged the appveyor branch with a fast-forward merge,
so two branches were referring to the same head commit, and appveyor
showed up in git show-ref above master, so that's what get_branch_name()
would return.