bin/summary --html -o file.html -v [-v]
Now you can see the information in plain text and also write the HTML
report in the same run.

This would be more useful if I still had a Jenkins, but it comes in
handy during manual testing: it's less boring to wait while all the data
is being collected if you get to see the data printed as it's being
collected.